Feature/implement new models

+++ src/main/java/com/erenberik/flightsearchapi/repository/FlightRepository.java
@@ -4,6 +4,6 @@ import com.erenberik.flightsearchapi.model.Flight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface FlightRepository extends JpaRepository<Flight,Integer> {
+public interface FlightRepository extends JpaRepository<Flight,Long> {
 
 }
